+++ example/.terraform/plugins/linux_amd64/lock.json
@@ -1,3 +1,3 @@
 {
-  "graylog": "934afd358d904f13cbdf1dcef4f7441675bacd1a4705cc40dae0dc67b0bd81fb"
+  "graylog": "1ea9b7184a4717e818171e396d7b44e8f180777c21cd527f32db9e0a946df30c"
 }
