+++ example/terraform.tfstate.backup
@@ -1,7 +1,7 @@
 {
     "version": 3,
     "terraform_version": "0.10.2",
-    "serial": 4,
+    "serial": 5,
     "lineage": "959f9d94-5f98-4d50-a64c-f5e6d88c9a9f",
     "modules": [
         {
@@ -14,17 +14,17 @@
                     "type": "graylog_input",
                     "depends_on": [],
                     "primary": {
-                        "id": "59f3aba22ab79c0001c94691",
+                        "id": "59f3b0732ab79c0001c94c56",
                         "attributes": {
                             "gelf_udp.#": "1",
                             "gelf_udp.2558345342.bind_address": "0.0.0.0",
                             "gelf_udp.2558345342.port": "22201",
                             "global": "true",
-                            "id": "59f3aba22ab79c0001c94691",
+                            "id": "59f3b0732ab79c0001c94c56",
                             "title": "GELF UDP"
                         },
                         "meta": {},
-                        "tainted": true
+                        "tainted": false
                     },
                     "deposed": [],
                     "provider": ""
